requirements.txt
openai==0.27.0
python-dotenv==1.0.0
requests==2.26.0
termcolor==1.1.0
langchain==0.0.136